# docs/api-openapi.yaml
openapi: 3.0.3
info:
  title: GPU Cloud Service API
  description: OpenAI-compatible inference and API key management for the GPU Cloud PaaS.
  version: 1.0.0

servers:
  - url: http://localhost:3001
    description: Local development

components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: Use your API key (e.g. gpu_...) or admin key.
  security:
    - ApiKeyAuth: []

paths:
  /health:
    get:
      summary: Health check
      operationId: getHealth
      security: []
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: ok }
                  timestamp: { type: string, format: date-time }

  /metrics:
    get:
      summary: Basic metrics (workers, queue depth)
      operationId: getMetrics
      security: []
      responses:
        '200':
          description: Metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  workers: { type: integer }
                  queueDepth: { type: integer }

  /api/keys:
    get:
      summary: List API keys
      operationId: listKeys
      responses:
        '200':
          description: List of keys (id, name, prefix, createdAt, lastUsedAt)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        name: { type: string }
                        prefix: { type: string }
                        createdAt: { type: string, format: date-time }
                        lastUsedAt: { type: string, format: date-time, nullable: true }
    post:
      summary: Create API key
      operationId: createKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, description: Key label }
      responses:
        '201':
          description: Key created; full key returned only once
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id: { type: string }
                      key: { type: string }
                      name: { type: string }
                      prefix: { type: string }
                      createdAt: { type: string }
        '400':
          description: Validation error

  /api/keys/{id}:
    delete:
      summary: Delete API key
      operationId: deleteKey
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Key deleted
        '404':
          description: Key not found

  /api/usage:
    get:
      summary: Usage summary
      operationId: getUsageSummary
      parameters:
        - name: keyId
          in: query
          schema: { type: string }
        - name: since
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Aggregated usage (totalRequests, totalTokens, etc.)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      summary:
                        type: object
                        properties:
                          totalRequests: { type: integer }
                          totalPromptTokens: { type: integer }
                          totalCompletionTokens: { type: integer }
                          totalTokens: { type: integer }

  /api/usage/history:
    get:
      summary: Usage history
      operationId: getUsageHistory
      parameters:
        - name: keyId
          in: query
          schema: { type: string }
        - name: since
          in: query
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        '200':
          description: Recent usage records

  /api/models:
    get:
      summary: List models and deployments
      operationId: listModels
      responses:
        '200':
          description: Available models and deployment list

  /api/deployments:
    get:
      summary: List deployments
      operationId: listDeployments
      responses:
        '200':
          description: Deployed models with status
    post:
      summary: Deploy a model (simulated)
      operationId: createDeployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
      responses:
        '201':
          description: Deployment created

  /v1/chat/completions:
    post:
      summary: Create chat completion (OpenAI-compatible)
      description: Supports both streaming and non-streaming. Uses simulated GPU workers.
      operationId: createChatCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model, messages]
              properties:
                model: { type: string }
                messages:
                  type: array
                  items:
                    type: object
                    required: [role, content]
                    properties:
                      role: { type: string, enum: [system, user, assistant] }
                      content: { type: string }
                stream: { type: boolean, default: false }
                max_tokens: { type: integer, default: 256 }
                temperature: { type: number, default: 0.7 }
      responses:
        '200':
          description: Non-streaming; returns full completion. Streaming returns SSE.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
                  object: { type: string, example: chat.completion }
                  created: { type: integer }
                  model: { type: string }
                  choices:
                    type: array
                    items:
                      type: object
                      properties:
                        index: { type: integer }
                        message: { type: object }
                        finish_reason: { type: string }
                  usage:
                    type: object
                    properties:
                      prompt_tokens: { type: integer }
                      completion_tokens: { type: integer }
                      total_tokens: { type: integer }
            text/event-stream:
              description: When stream=true, Server-Sent Events with chat.completion.chunk
        '400':
          description: Invalid request (e.g. missing messages)
        '401':
          description: Invalid or missing API key
        '429':
          description: Rate limit exceeded